после слияния

+++ main.py
@@ -4,6 +4,8 @@ from vk_api.longpoll import VkLongPoll, VkEventType
 from key_words import *
 from bd_users import DB, UserModel
 from vk_metods import *
+from vk_api.utils import get_random_id
+import wikipedia
 import logging
 
 token = "f9a38be71df81831bd854f44bf26613d5f91dd6edeec8164f45261f5def3aba84d019b14ef3b914c869a2"
@@ -11,12 +13,14 @@ token = "f9a38be71df81831bd854f44bf26613d5f91dd6edeec8164f45261f5def3aba84d019b1
 vk = vk_api.VkApi(token=token)
 longpoll = VkLongPoll(vk)
 db = DB()
+wikipedia.set_lang("RU")
 logging.basicConfig(filename='info.log', 
                     level=logging.INFO,
                     format='%(asctime)s %(levelname)s %(name)s %(message)s')
 
 
 def main():
+    wiki = False
     UserModel(db.get_connection()).init_table()
     user_model = UserModel(db.get_connection())
     for event in longpoll.listen():
@@ -34,20 +38,25 @@ def main():
             for fraze in TALK:
                 if fraze in all_request:
                     pass
-            for word in request:
-                if word in HELLO:
-                    response.append(random.choice(HELLO) + ', ' + info['first_name'])
-                if word == "клава":
-                    keyboard_vk(event.user_id)
-                if word == "вики":
-                    wiki = True
-                    response.append("Введите запрос")
-                if word == "пока" and len(response) == 0:
-                    response.append("Пока")
-                    user_model.delete_user(event.user_id)
-                    logging.info('Завершение диалога с {}(id={})'.format(info['first_name'], info['id']))
-                if len(response) == 0:
-                    response.append("Не понялa вашего ответа..")
+            if wiki:
+                response.append('Вот что я нашёл: \n' + str(wikipedia.summary(event.text)))
+                wiki = False
+            else:
+                for word in request:
+                    if word in HELLO:
+                        response.append(random.choice(HELLO) + ', ' + info['first_name'])
+                    if word == "клава":
+                        keyboard_vk(event.user_id)
+                        response.append("..")
+                    if word == "вики":
+                        wiki = True
+                        response.append("Введите запрос")
+                    if word == "пока" and len(response) == 0:
+                        response.append("Пока")
+                        user_model.delete_user(event.user_id)
+                        logging.info('Завершение диалога с {}(id={})'.format(info['first_name'], info['id']))
+                    if len(response) == 0:
+                        response.append("Не понялa вашего ответа..")
             response = map(lambda x: x[0].upper() + x[1:] + '. ', response)
             response = ''.join(response)
             write_msg(event.user_id, response)
